+++ Sms/Cpu/Instructions/Arithmetic8Bit/ADD_A__IX_d_.cs
@@ -8,7 +8,7 @@
 
         protected override void InnerExecute(byte opCode)
         {
-            var d = Z80.Memory[Z80.Registers.PC++];
+            var d = (sbyte)Z80.Memory[Z80.Registers.PC++];
             var value = Z80.Memory[(ushort)(Z80.Registers.IX + d)];
 
             Z80.Alu.Add(value);
