+++ Sms/Cpu/Instructions/BitSetResetAndTest/SET_b__IX_d_.cs
@@ -16,7 +16,7 @@
         protected override void InnerExecute(byte opCode)
         {
             var b = (opCode & 0b00111000) >> 3;
-            var d = Z80.Memory[(ushort)(Z80.Registers.PC - 2)];
+            var d = (sbyte)Z80.Memory[(ushort)(Z80.Registers.PC - 2)];
 
             var value = Z80.Memory[(ushort)(Z80.Registers.IX + d)];
             value = value.SetBit(b);
